Restructure project, add files for storage to implement CRUD operation on MongoDB.
