+++ src/Services/Catalog/go.mod
@@ -6,6 +6,7 @@ require (
 	github.com/gin-contrib/cors v1.4.0
 	github.com/gin-gonic/gin v1.9.0
 	github.com/go-playground/validator v9.31.0+incompatible
+	github.com/huavanthong/microservice-golang/product-api-v3 v0.0.0-20230218084159-b142ef4fac48
 	github.com/natefinch/lumberjack v2.0.0+incompatible
 	github.com/sirupsen/logrus v1.6.0
 	github.com/spf13/viper v1.12.0
@@ -17,4 +18,3 @@ require (
 	google.golang.org/genproto v0.0.0-20220519153652-3a47de7e79bd
 	gopkg.in/natefinch/lumberjack.v2 v2.2.1 // indirect
 )
-
