+++ src/Services/Catalog/internal/infrastructure/storage/mongodb/category.mongo.go
@@ -1,4 +1,4 @@
-package repositories
+package storage
 
 import (
 	"context"
