+++ src/Services/Catalog/internal/infrastructure/storage/mongodb/product.mongo.go
@@ -1,4 +1,4 @@
-package repositories
+package mongodb
 
 import (
 	"context"
